requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.3
django-environ==0.11.2
django-rest-knox==5.0.2
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
dynaconf==3.2.6
gunicorn==23.0.0
hvac==2.3.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
knox==0.1.14
loguru==0.7.2
packaging==24.2
psycopg2==2.9.10
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
shortuuid==1.0.13
sqlparse==0.5.2
uritemplate==4.1.1
urllib3==2.2.3
validators==0.34.0
